TopBarActions component created
TopBarActions component created and fixed some small issues
